fix MANDATORY_FIELDS comparisons (need to exclude -1)

+++ eeschema/class_libentry.cpp
@@ -1099,7 +1099,7 @@ bool LIB_PART::LoadField( LINE_READER& aLineReader, wxString& aErrorMsg )
         return false;
     }
 
-    if( field->GetId() < MANDATORY_FIELDS )
+    if( (unsigned) field->GetId() < MANDATORY_FIELDS )
     {
         LIB_FIELD* fixedField = GetField( field->GetId() );
